Updates according to feedback
