--- src/main/java/org/craftercms/studio/api/v2/service/translation/internal/TranslationServiceInternal.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (C) 2007-2020 Crafter Software Corporation. All Rights Reserved.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as published by
- * the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.craftercms.studio.api.v2.service.translation.internal;
-
-import org.craftercms.studio.api.v1.to.TranslationConfigTo;
-
-/**
- * Defines all translation related operations
- *
- * @author joseross
- * @since 3.2.0
- */
-public interface TranslationServiceInternal {
-
-    /**
-     * Get the translation configuration for a given site
-     * @param siteId the id of the site
-     * @return the configuration
-     */
-    TranslationConfigTo getConfig(String siteId);
-
-}
