+++ src/main/resources/crafter/studio/studio-services-context.xml
@@ -1325,18 +1325,5 @@
         <constructor-arg ref="encryptionServiceInternal"/>
     </bean>
 
-    <!-- ////////////////////////////////////// -->
-    <!--             Translation                -->
-    <!-- ////////////////////////////////////// -->
-
-    <bean id="translationServiceInternal"
-          class="org.craftercms.studio.impl.v2.service.translation.internal.TranslationServiceInternalImpl">
-        <constructor-arg ref="cstudioServicesConfig"/>
-    </bean>
-
-    <bean id="translationService" class="org.craftercms.studio.impl.v2.service.translation.TranslationServiceImpl">
-        <constructor-arg ref="translationServiceInternal"/>
-    </bean>
-
     <import resource="classpath:crafter/studio/studio-security-context.xml"/>
 </beans>
